fix(resize): component doesn't recalculate wrappers reliably
